Add monitor generator bind
