+++ magma/__init__.py
@@ -83,3 +83,4 @@ def get_family():
     return _Family_
 
 from .generator import Generator
+from .monitor import MonitorIO, MonitorGenerator
